SI-Reference-Point-2023-14 Addition of other JSON-LD outputs (via content negotiation) for other ...
Added JSON-LD output via content negotiation for constants. In the process I rewrote the SPARQL query to get the data out of the constant.ttl file and this query can also be reused for the /constants endpoint.  However, the SPARQL query does not return the hasUnitElement data as I need to get help from Jean-Laurent with that.  The output template file for constant/Avogadro/ was updated to visualize all the fields of data that are available for constants.  An example of how to get JSON-LD out is to use the following command line request using curl.  curl http://localhost:8000/constant/Avogadro/ -H "Accept: application/ld+json" -o 'avogadro.jsonld'