SI-Reference-Point-2023-14 Addition of other JSON-LD outputs (via content negotiation) for other ...
Added prefixes to JSON-LD at /si-prefixes/.  The response is language independent with both the English and French being return in the file.  Added urls to pages for the individual prefixes if we decide to implement them. To get the JSON-LD use the following curl command on the command line: curl http://localhost:8000/si-prefixes/ -H "Accept: application/ld+json" -o 'prefixes.jsonld'