SI-Reference-Point-2023-14 Addition of other JSON-LD outputs (via content negotiation) for other ...
Added quantities to JSON-LD at /quantities/.  The response can be in English or French. Example download can be requested on the command line using the following for the English version.  Note that the URL must be enclosed in quotes '' so the GET parameter lang can be used on the command line.

curl 'http://localhost:8000/quantities/?lang=en' -H "Accept: application/ld+json" -o 'quantities.jsonld'