Deal with compound units in quantities.yaml
Update of quantities.yaml by JLH

I saw that Units_Abox.py already add a function to build compound units
(used for filling the inOtherUnits and inBaseSIUnits attributes), so I
reconnected to that.

Doing so I realize we invented yet another way to encode compound units.

E.g for acceleration the function transform_to_graph expects a dict
looking like :

{"mult":[{"exp":['metre', 1]}, {"exp":['second', -2]}

whereas we decided to put this in the yaml

 Unit:
  - [metre, 1]
  - [second, -2]

(and it does extract the dict from a string).

Maybe we could just choose to use the dict notation directlyt in the
yaml, or the string notation. Then again it is a matter of saving 2 lines of code,
so maybe also not necessary.