Add concept of restricted prefix (+ add hectare unit)
This commit introduces :

prefixRestriction : expected to point to a boolean, indicates that not
all prefixes can be used in conjunction with this unit. Only present
when "true" (we might want to make this mandatory and pointing to
"false" byu default, but it is not the case yet).

hasAuthorizedPrefix : a URI pointing to a prefix that is authorized to
be used. E.g. Bel can only have deci-. In case multiple prefixes are
allowed, add triples with this predicate.

Separating the two allows to explicitely disallow any prefix (by issuing
the prefixRestriction = true, and issuing no hasAuthorizedPrefix).

Also, hectare has been added to the list of non-si units